+++ public/js/src/module/translator.js
@@ -68,7 +68,7 @@ const t = ( key, options ) => i18next.t( key, options );
  * Note, this does not work if there is translation context (i.e. options for t(key, options) call).
  *
  * @param {Element} container - The element to localize.
- * @param { string } [lng] - The 2-character IANA subtag.
+ * @param { string } [lng] - The 2-or-3-character IANA subtag.
  */
 const localize = ( container, lng ) => {
     const cache = {};
@@ -139,10 +139,8 @@ const getCurrentUiLanguage = () => {
     return matches.length ? matches[ 0 ] : null;
 };
 
-const getDesiredLanguage = () => {
-    // Here we are essential duplicating the LanguageDetector module, which is not great, but very simple in our case.
-    const src = settings.languageOverrideParameter ? settings.languageOverrideParameter.value : navigator.language;
-    const matches = src.match( LANGEXTRACT );
+const getBrowserLanguage = ( ) => {
+    const matches = navigator.language.match( LANGEXTRACT );
 
     return matches.length ? matches[ 0 ] : null;
 };
@@ -153,7 +151,7 @@ export {
     localize,
     loadTranslation,
     getCurrentUiLanguage,
-    getDesiredLanguage
+    getBrowserLanguage
 };
 
 /**